Added infobundle methods to test bundle

+++ src/slogo/events/EnvironmentRecord.java
@@ -5,5 +5,5 @@ import slogo.model.Palette;
 public record EnvironmentRecord(
     Palette colors,
     int currPenColor, int currShape,
-    int currBGColor, int currPenSize) {
+    int currBGColor, double currPenSize) {
 }
